+++ src/Pixel.h
@@ -22,6 +22,6 @@ struct Pixel {
 void GetPixList(const Image<double>& im, std::vector<Pixel>& pix,
     const Position cen, double sky, double noise, double gain,
     const Image<double>* wt_im, const Transformation& trans,
-    double aperture, int& flag);
+    double aperture, long& flag);
 
 #endif
